Parameter Shadow Warnings from CEditable Item

+++ apl/CFGC/CFGC.cpp
@@ -99,17 +99,14 @@ uint16_t CFGC::GetPasswords(uint8_t u8PasswordNo)
     switch(u8PasswordNo)
     {
         case 0:
-            return ((_stCfgc.u8PIN1PasswordChar0 * 1000U) + (_stCfgc.u8PIN1PasswordChar1 * 100U) +
-                    (_stCfgc.u8PIN1PasswordChar2 * 10U) + (_stCfgc.u8PIN1PasswordChar3));
+            return (uint16_t)((_stCfgc.u8PIN1PasswordChar0 * 1000U) + (_stCfgc.u8PIN1PasswordChar1 * 100U) +(_stCfgc.u8PIN1PasswordChar2 * 10U) + (_stCfgc.u8PIN1PasswordChar3));
         break;
         case 1:
-            return ((_stCfgc.u8PIN2PasswordChar0 * 1000U) + (_stCfgc.u8PIN2PasswordChar1 * 100U) +
-                                (_stCfgc.u8PIN2PasswordChar2 * 10U) + (_stCfgc.u8PIN2PasswordChar3));
+            return (uint16_t)((_stCfgc.u8PIN2PasswordChar0 * 1000U) + (_stCfgc.u8PIN2PasswordChar1 * 100U) +(_stCfgc.u8PIN2PasswordChar2 * 10U) + (_stCfgc.u8PIN2PasswordChar3));
         break;
 
         default:
-            return ((_stCfgc.u8PIN1PasswordChar0 * 1000U) + (_stCfgc.u8PIN1PasswordChar1 * 100U) +
-                                (_stCfgc.u8PIN1PasswordChar2 * 10U) + (_stCfgc.u8PIN1PasswordChar3));
+            return (uint16_t)((_stCfgc.u8PIN1PasswordChar0 * 1000U) + (_stCfgc.u8PIN1PasswordChar1 * 100U) +(_stCfgc.u8PIN1PasswordChar2 * 10U) + (_stCfgc.u8PIN1PasswordChar3));
             break;
     }
